ENH: Replacing eta_squared matlab with python version
Notes - I think there is a bug in the matlab version caused
by integer arithmetic. When the input images are of different
pixel types the stacking operation seems to default to the
lower precision type (typically int16) and this gets
used for arithmetic later on.

The matlab version is set up to take two inputs, but the code is
more general, creating the correlation matrix for an arbitary
size array stack. However only a single matrix entry is
returned.

The python version is only computing the single matrix entry,
corresponding to the pairwise eta.